+++ src/routing/ors_wrapper.h
@@ -23,21 +23,8 @@ private:
                       const std::vector<Location>& locs,
                       const std::string& service) const override;
 
-  bool
-  duration_value_is_null(const rapidjson::Value& matrix_entry) const override;
-
-  bool
-  distance_value_is_null(const rapidjson::Value& matrix_entry) const override;
-
-  UserDuration
-  get_duration_value(const rapidjson::Value& matrix_entry) const override;
-
-  UserDistance
-  get_distance_value(const rapidjson::Value& matrix_entry) const override;
-
-  unsigned get_legs_number(const rapidjson::Value& result) const override;
-
-  std::string get_geometry(rapidjson::Value& result) const override;
+  const rapidjson::Value&
+  get_legs(const rapidjson::Value& result) const override;
 
 public:
   OrsWrapper(const std::string& profile, const Server& server);